requirements.txt


Django==4.0.4
django-debug-toolbar==3.4.0
django-cors-headers==3.11.0
django-environ==0.8.1
djangorestframework==3.13.1
drf-nested-routers==0.93.4
